+++ android/src/main/java/io/amarcruz/rnsettings/RNSettingsPackage.java
@@ -18,12 +18,9 @@ public class RNSettingsPackage implements ReactPackage {
     }
 
     @Override
-    public List<NativeModule> createNativeModules(
-            ReactApplicationContext reactContext) {
+    public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
         List<NativeModule> modules = new ArrayList<>();
-
         modules.add(new RNSettingsModule(reactContext));
-
         return modules;
     }