+++ index.d.ts
@@ -5,8 +5,8 @@ declare module "react-native-cross-settings" {
   export interface SettingsStatic {
     get(key: string): SettingsTypes;
     set(settings: { [key: string]: SettingsTypes }): void;
-    watchKeys(keys: string | string[], callback: () => any): any
-    clearWatch(watchId: any): void;
+    watchKeys(keys: string | string[], callback: () => void): number
+    clearWatch(watchId: number): void;
   }
 
   const Settings: SettingsStatic;