+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2018-10-09
+
+This is the first version for Gradle 4.4
+
+### Added
+- Runtime check in debug mode for the type of values to store.
+
+### Changed
+- The id of `watchKeys` and `clearWatch` is number, the callback returns void.
+
 ## [1.0.2] - 2018-09-17
 ### Changed
 - The changelog follows the format on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).